Extract Dry::Initializer singleton methods to DSL module
